feat: add internal JWT authentication and user registration views
