+++ .env.example
@@ -7,3 +7,4 @@ DB_HOST=db
 DB_PORT=5432
 ALLOWED_HOSTS=0.0.0.0,localhost,render.com
 INTERNAL_JWT_SECRET_KEY=your-internal-jwt-secret-key
+INTERNAL_JWT_ALLOWED_SERVICES=sugarfoot,koda,gary
